+++ internal/util/misc.go
@@ -47,7 +47,7 @@ func ExtractBaseHost(rawURL string) (string, error) {
 	return parts[0], nil
 }
 
-func ExceedsMaxFileSize(fileSize int32) bool {
+func ExceedsMaxFileSize(fileSize int64) bool {
 	return fileSize > config.Env.MaxFileSize
 }
